NetworkBrowser: Fixes Issue with "Empty Folders"
E.g.
- Mounted DVDs on PC via CIFS Share
- Mounted DVD-Subfolders on PC via CIFS-Share
- NAS-Shares via CIFS

It's a issue introduced with Openpli 2.1 assumed together with some
LFS-changes. So this this the best and easiest place to quickly fix
this issue directly in Plugin to reach the same behaviour for CIFS-Share
Mounting like before with OpenPli 2.0. In Openpli 2.0 "noserverino" was
given as default, with openli 2.1 "serverino" was given by default, but
"serverino" actually causes the empty Folders issues for CIFS-Shares.