Network browser cleanup (#325)
* [Networkbrowser] Cleanup of network address handling

Put common code to generate the string form of the network address
in x.x.x.x/netmaskWidth format into a new method makeIpStr().

Check whether both the returned IP address and netmask for error
returns from iNetwork.getAdapterAttribute().

If a network attribute can't be found by iNetwork.getAdapterAttribute()
it returns None, not an empty list or tuple. Change the tests for
the error return from

if len(retval):

to simply

if retval:

so that either a None or an empty return value will fail the test
correctly, and add tests on the length of the IP address
and netmask to be equal to 4.

Simplify the generation of the netmask bitcount by removing the
unnecesary call of formatIp().

Simplify the generation of the network IP address string by using
a list comprehension instead of enumerating all the IP address
and netmask parts.

Simplify getNetworkIPs() by simply returning the return
value of netscan.netzInfo() rather than appending. Also
avoid a potential crash if makeIpStr() returns None.

* [Networkbrowser] Prevent near-infinite loop & crash

Check whether iNetwork.getAdapterAttribute() returns an all-zero
IP address or an all-zero netmask, and give an error popup in either
case.

This prevents nmap from enumerating the whole IPv4 address space
if the network scan is run when there is no network interface
available.

If this happens, nmap also returns one line of XML data for each
IP address scanned, which is accumulated in the Console containing
the run of nmap. This eventually leads to memory being completely
consumed and a crash of the UI.